Move asset-url code back
